Backend model work (#6)
* That's where my commit went

* A lot of backend changes involving a new update model, and grid system

I didn't have enough patience with angular to try and make the front end work as well yet. But I started a loop that should theoretically give you access to all the entities that should be drawn

* Fixed serialization loop errors
and a bunch of null references

* That's where my commit went

* A lot of backend changes involving a new update model, and grid system

I didn't have enough patience with angular to try and make the front end work as well yet. But I started a loop that should theoretically give you access to all the entities that should be drawn

* Fixed serialization loop errors
and a bunch of null references

* look at branch name.  also added line so we know what direction its moving in.  I know some of this might change with backend work, like using velocity over actual coords.

* More serialization fixes, and added up to 16 objects being drawn

* Added a green box, that's supposed to turn red if connection is lost

Also, yellow line points at the green box, so that you can always return to middle

Added a bunch of try/catches

And handleded player disconnection

Made things a bit more stable per connection in a concurrent environment

Added the option to draw a lot more players